if using --pretend, create new folder adding _v1 to name to avoid mistakes and print commands

+++ WMass/python/plotter/w-helicity-13TeV/skims.py
@@ -51,13 +51,21 @@ if __name__ == "__main__":
             os.makedirs(outputDirSkims)
             os.makedirs(outputDirFSkims)
         else:
-            print "The skim output dir ",outputDirSkims," exists. Will remove it and substitute with new one. \nDo you agree?[y/N]\n"
-            if raw_input()!='y':
-                print 'Aborting'
-                exit()
-            os.system("rm -rf "+outputDirSkims)
-            os.makedirs(outputDirSkims)
-            os.makedirs(outputDirFSkims)
+            if options.pretend:
+                print "The skim output dir ",outputDirSkims," exists, but you passed option --pretend"
+                print "I guess you just want to print commands to test behaviour or resubmit few jobs"
+                print "I will add suffix _v1 to the output folder to let you avoid making mistakes and overwriting things"            
+                outputDirSkims = outputDirSkims + "_v1"
+                outputDirFSkims = outputDirSkims+"/friends"
+                print "New output dir is " + outputDirSkims
+            else:
+                print "The skim output dir ",outputDirSkims," exists. Will remove it and substitute with new one. \nDo you agree?[y/N]\n"
+                if raw_input()!='y':
+                    print 'Aborting'
+                    exit()
+                os.system("rm -rf "+outputDirSkims)
+                os.makedirs(outputDirSkims)
+                os.makedirs(outputDirFSkims)
         os.system('cp {vf} {od}'.format(od=outputDirSkims,vf=options.varfile))
         os.system('cp {sf} {od}'.format(od=outputDirSkims,sf=args[1])) ## this should work??
     else: print "Make only the friend trees in dir ",outputDirFSkims
